requirements.txt

# requirements.txt

numpy
himage
pycryptodome
pillow
